Found a work around for the disposing of the OMScontext instance. The using block should uphold that data in case it is or isn't created. Also, should be switching to binding source, haven't learned about it yet, but could be useful. Also, changing up the database again. There are some issues with Contacts not linking back to the Events, which would be the organizer of the event. At least one contact should be necessary as a point of contacts, and further which organization that seup the event. Under the relationships, we have that "Contacts can be associated with muliple events (many-to-many relationship)." So, figured out how to use the conventions for relationship discovery. Have to add the = new List<XModel>();. Which I'm assuming is it create a list for the storage of the appropriate references for the associated type.
Smaller fix, "VenueID" changed to "VenuesID" to be more consistent with the cname of the class.

Going to add the migration, and update the database.
That failed. Chaning it made some changes to constraints and columns could no longer be null. Decided to drop all the migrations and go ahead and restart the database.

Trying to add mirations and update the database again...